update(components): refactor product-status-dropdown, product-category-search, product-tax-dropdown
