+++ stocky-web/src/app/routes/products/_components/product-tax/product-tax-add/product-tax-add.component.html
@@ -48,7 +48,6 @@
                    [props]="{tableSize: 'small'}"
                    [rowTemplates]="[titleTmpl, percentTmpl, descTmpl, actionTmpl]"
         ></app-table>
-
     </div>
 </ng-container>
